+++ src/include/sGameServices.h
@@ -5,6 +5,7 @@ class cGameObjectContext;
 class cInfoContext;
 class cGameSettings;
 class cLog;
+class cStructureUtils;
 
 struct sGameServices
 {
@@ -12,5 +13,6 @@ struct sGameServices
     cGameObjectContext *objects = nullptr;
     cInfoContext *info = nullptr;
     cGameSettings *settings = nullptr;
+    cStructureUtils *structureUtils = nullptr;
     cLog *m_log = nullptr;
 };
